% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa_helper.R
\name{create_disposition_view}
\alias{create_disposition_view}
\title{Creates a helpful view from student Pisa data.}
\usage{
create_disposition_view(students)
}
\arguments{
\item{students}{A data.table modified from student2012}
}
\value{
A data.table prepared to be passed into a GBM.
}
\description{
\code{create_disposition_view} Creates a view for student disposition
}
\examples{
create_disposition_view(student2012)
}
