% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pisa_helper.R
\name{train_gbm}
\alias{train_gbm}
\title{Trains a model for student dispositions effects on math scores.}
\usage{
train_gbm(students)
}
\arguments{
\item{students}{A data.table modified to contain student disposition columns}
}
\value{
A trained supervised model created from h2o
}
\description{
\code{train_gbm} Uses a gbm to determine a model for math scores and student dispositon.
}
\examples{
train_gbm(australian_disposition)
}
